Fix context menus not working in TemporaryContainer grouped container, closes #76
The solution makes context menus operate on CTS containers instead of
browser containers, effectively giving context menu context of a
particular container.

Not only does it make support of TemporaryContainer possible, but it
also allows Pinned Tabs container Context Menu to work on pinned tabs.

The commit also introduces a concept of container id, it's cookieStoreId
for non special containers and preset ids that can be found in constants.js
for other containers.